intercept: search closest approaches of orbits
A search algorithm for closest approaches between two satellites in
orbit.

The algorithm can be used to detect intercept and rendez-vous of two
satellites or find sphere of influence transitions on lunar and
interplanetary trajectories.

The algorithm works in three steps: geometry prefilter, time prefilter
and minimization.

The geometry prefilter computes ranges of true anomaly where the
satellites may be closer than a set threshold.

The time prefilter computes time intervals where both satellites are in
the range specified by the geometry prefilter.

The minimization is a root finding algorithm that evaluates positions
and velocities of the satellites, taking advantage of a priori knowledge
of the physics of two body orbital motion to reduce the number of steps
required.

This patch is squashed from more than 50 commits written over a period
of 20 months.