intercept: test params in test, fix infinite loop
intercept_test eccentric anomaly generation had an operator precedence
bug that caused E1 and E2 to always get value pi or pi/2, drastically
reducing test coverage.

This was discovered because of a new compiler warning and fixed with a
pair of parenthesis.

This change revealed a bug in intercept_times, causing an infinite loop
at very high eccentricity because it was considered "open" by
conic_closed but not an ellipse by conic_elliptic.

Use relative error to check for intercept times.

There are failing tests after this commit, only 99% of intercept_test
passes.